Added Puntos Layout
