+++ Utils/constants.ts
@@ -2,4 +2,5 @@ export const BASE_URL = 'https://6222994f666291106a29f999.mockapi.io/api/v1';
 
 export const COLORS = {
   BLACK: '#020202',
+  GRAY: '#9B9898',
 };
